feat: 닫기 버튼이 있는 모달 구현
